chore(refactor): move NamespaceFilter type to src/config/types (refs #47)
